Make the generate validator use the test syntax
Runs all generators, even if one fails
Updated travis.sh
